(major) move away from kristijanhusak/orgmode.nvim
defer learning orgmode for a later date, and opt for the more palatable
vimwiki/vimwiki. Not that orgmode is insufficient, in fact it is too
powerful for my current requirements! vimwiki is /comfy/ enough :)

TODO: remember to automate the uninstallation of the treesitter
grammars. Currently have to uninstall the org TSgrammar since it doesn't
have first class support.

- remove akinsho/org-bullets and config since it is no longer required
- remove michaelb/sniprun since it is not useful for the workflow (i.e.
  Jupyter notebooks are unavoidable!)