feat(emacs): setup org-mode, org-roam, org-roam-ui
Pretty big change, with configurations for org-mode, org-roam and
org-roam-ui.
NOTE: See the code for the related shortcuts, and which folders
are expected to be created for storing .org, and org-roam files.

Additional changes are as follows:
- set up ligature.el, which is (now) available on MELPA.
- add configuration for org-bullets, for those sweet looking org-mode
lists
- customize the font faces for various things. NOTE: you will need the
Indian Font Foundry's General Sans, and Chubbo for this to render
as intended. Put these in /usr/share/fonts/non-free and you're golden.
- couple of shortcuts added to enable telescope-like file finding and
navigation
- line numbers were disabled for some modes (see code for exact config)